+++ src/main/java/com/example/seoulpublicdata2025backend/domain/kakaoSocialLogin/controller/KakaoLoginController.java
@@ -1,16 +1,13 @@
 package com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.controller;
 
 import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.dto.KakaoUserInfoResponseDto;
-import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.service.KakaoService;
+import com.example.seoulpublicdata2025backend.domain.kakaoSocialLogin.service.KakaoServiceImpl;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
 
 @Slf4j
 //@RestController
@@ -19,7 +16,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("")
 public class KakaoLoginController {
 
-    private final KakaoService kakaoService;
+    private final KakaoServiceImpl kakaoService;
 
     @GetMapping("/auth/login/kakao")
     public String KakaoLogin(@RequestParam("code") String code) {
@@ -33,8 +30,11 @@ public class KakaoLoginController {
         // 여기에 서버 사용자 로그인(인증) 또는 회원가입 로직 추가
 
         Long kakaoId = userInfo.getId();
-        System.out.println(kakaoId);
+        String profileUrl = profile.getProfileImageUrl();
+
+//        System.out.println(kakaoId);
+//        System.out.println(profile.profileImageUrl);
 //        return new ResponseEntity<>(HttpStatus.OK);
-        return "redirect:/signup?kakaoId=" + kakaoId;
+        return "redirect:/signup?kakaoId=" + kakaoId + "&profileUrl=" + profileUrl;
     }
 }
