+++ src/main/java/com/example/seoulpublicdata2025backend/domain/kakaoSocialLogin/controller/SignupPageController.java
@@ -11,8 +11,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 public class SignupPageController {
 
     @GetMapping
-    public String signupPage(@RequestParam("kakaoId") Long kakaoId, Model model) {
+    public String signupPage(@RequestParam("kakaoId") Long kakaoId, @RequestParam("profileUrl") String profileUrl, Model model) {
         model.addAttribute("kakaoId", kakaoId);
+        model.addAttribute("profileUrl", profileUrl);
         return "signup";
     }
 }
